% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq2.R
\name{chisq2.test}
\alias{chisq2.test}
\title{Pearson's Chi-squared Test for Count Data}
\usage{
chisq2.test(obs, p.esp, npar = NULL, grouping = TRUE)
}
\arguments{
\item{obs}{blaba}

\item{p.esp}{bla.}

\item{npar}{blabla}

\item{grouping}{blabla}
}
\value{
test balb alb al a.
}
\description{
bla bla bla
}
