% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPD.R
\docType{package}
\name{CPD-package}
\alias{CPD}
\alias{CPD-package}
\title{CRISPR/Cas9 gRNA designer}
\description{
CPD is a R/Bioconductor package for CRISPR/Cas9 gRNA design. 
The input is GenomicRanges object. 
Output will be multiple excel sheet of gRNAs and off targets.
}
