% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpd.R
\docType{package}
\name{cpd}
\alias{cpd}
\alias{cpd-package}
\alias{cpd-package}
\title{Pmf for the CTP distribution}
\description{
Complex Pearson Distribution ... .
}
\details{
The only function you're likely to need from roxygen2 is [roxygenize()].
Otherwise refer to the vignettes to see how to format the documentation.
}
\author{
\strong{Maintainer}: Silverio Vilchez-Lopez \email{svilchez@ujaen.es}

Authors:
\itemize{
  \item Maria Jose Olmo-Jimenez
}

}
\keyword{internal}
