% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dctp.R, R/pctp.R, R/qctp.R, R/rctp.R
\name{ctp}
\alias{ctp}
\alias{dctp}
\alias{pctp}
\alias{qctp}
\alias{rctp}
\title{The Complex Triparametric Pearson (CTP) Distribution}
\source{
Rodr\'iguez-Avi, J., S\'aez-Castillo, A. J., Conde-Sánchez, A. and Olmo-Jim\'enez, M. J. (2004). A triparametric discrete distribution with complex parameters. Statistical Papers, 45 (1), 81-95.
Olmo-Jim\'enez, M. J., Rodr\'iguez-Avi, J. and Cueva-López, V. (2018). A review of the CTP distribution: a comparison with other over- and underdispersed count data models. Journal of Statistical Computation and Simulation, 88 (14), 2684-2706.
}
\usage{
dctp(x, a, b, g)

pctp(x, a, b, g, lower.tail = TRUE)

qctp(p, a, b, g, lower.tail = TRUE)

rctp(n, a, b, g)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles}

\item{a}{parameter a (real)}

\item{b}{parameter b (real)}

\item{g}{parameter gamma (positive)}

\item{lower.tail}{A param g}

\item{p}{vector of (non-negative integer) probabilities}

\item{n}{number of observations. If length(n) > 1, the length is
taken to be the number required.}
}
\value{
The CTP distribution with parameters a, b and gamma has pmf
\eqn{f(x|a,b,\gamma)=f_0\frac{(a+ib)_x(a-ib)_x}{(\gamma)_x}\frac{1}{x!},\quad x=0,1,...}
with \eqn{f_0=\frac{\Gamma(\gamma-a-ib)\Gamma(\gamma-a+ib)}{\Gamma(\gamma)\Gamma(\gamma-2a)}} the normalizing constant.
If a=0 the CTP is a Complex Biparametric Pearson (CBP) distribution, so the pmf of the CBP distribution is obtained.
}
\description{
Probability mass function for the Complex Triparametric Pearson (CTP) distribution with parameters a, b and gamma.
}
\details{
bla blala lakj akj j fkjlh jakh bjksh jkh ajbh akjh kajsh jb ajbfd
  alskjfdaj sdfhlaskjhf lasdkjhf kjasdhfkjadshfkjahds fkjh asdkjfh asdjkhf
  kajf sdhlkasjdhf lkasdjhf lkasdjhf kasdjh flkasjh flksajdh flkasjhd jhasdf l
}
\examples{
#Ejemplo uso dctp
dctp(3,1,2,5)
dctp(c(3,4),1,2,5)


pctp(3,1,2,3)
pctp(c(3,4),1,2,3)
qctp(0.5,1,2,3)
qctp(c(.8,.9),1,2,3)
rctp(4,1,1,3)
}
