% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRes.R
\name{filterRes}
\alias{filterRes}
\title{filter offTargetAnalysis results}
\usage{
filterRes(results, outputDir, ...)
}
\arguments{
\item{results}{output of offTargetAnalysis}

\item{outputDir}{output folder of oofTargetAnalysis}

\item{...}{not used}
}
\value{
a sorted gRNAs table with efficay, off-target number, and unique restriction enzyme cut sites.
}
\description{
filter offTargetAnalysis results by sorting the results with best efficacy and shortest off target list.
}
