% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRNA.R
\name{gRNA}
\alias{gRNA}
\title{gRNA designer}
\usage{
gRNA(gr, upstream = 2000, downstream = 2000, anchor = c("TSS", "TES",
  "ALL", "onlyUpstream&Downstream"), species = c("mm10", "mm9", "hg19",
  "hg38", "danRer10"), ...)
}
\arguments{
\item{gr}{A \link[GenomicRanges:GRanges-class]{GRanges} object or coordinates files could be converted into GRanges.}

\item{upstream, downstream}{numeric(1).upstream or downstream of the coordinates.}

\item{anchor}{Anchor point of upstream and downstream. TSS: search gRNAs for promoter of given coordinates.}

\item{species}{character(1). Avaliable values: "mm9", "mm10", "hg19", "hg38", "danRer10"}

\item{...}{Parameters could be passed to \link[CRISPRseek:offTargetAnalysis]{offTargetAnalysis}. 
Default settings are: scoring.method = "CFDscore", annotatePaired = FALSE, max.mismatch=1}
}
\value{
invisible list of results of offTargetAnalysis
}
\description{
CRISPR/Cas9 gRNA designer.
}
\examples{
gr <- system.file('extdata', 'Sftpc.bed', package = 'CPD')
gRNA(gr, anchor="onlyUpstream&Downstream", species="mm10", 
       chromToSearch="chr14", outputDir="test")

}
