% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNgRNA.R
\name{topNgRNA}
\alias{topNgRNA}
\title{merge the results from multiple folders}
\usage{
topNgRNA(output = "topNgRNA.csv", path = ".",
  filename = "on.target.summary.csv", n = 2, cutoffEfficacy = 0,
  cutofftop5OfftargetTotalScore = 5, ...)
}
\arguments{
\item{output}{output file of summary file. Default is topNgRNA.csv.}

\item{path}{Path to be searched. Default is current folder.}

\item{filename}{file name to be searched. Default is on.target.summary.csv}

\item{n}{top N gRNAs per file. Default is 2.}

\item{cutoffEfficacy}{cutoff value of gRNA efficacy value. Default is 0}

\item{cutofftop5OfftargetTotalScore}{cutoff value of top5OfftargetTotalScore. Default is 5.}

\item{...}{not used.}
}
\value{
a dataframe
}
\description{
merge multiple on.target.summary.csv into one file
}
