% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.test2.R
\name{chisq.test2}
\alias{chisq.test2}
\title{Pearson's Chi-squared Test for Count Data}
\usage{
chisq.test2(obs, p.esp, npar = NULL, grouping = FALSE)
}
\arguments{
\item{obs}{a numeric vector with the counts}

\item{p.esp}{a numeric vector with the expected probabilities of the same length of \code{obs}. They must sum 1.}

\item{npar}{an integer specifying the number of parameters of the model. By default \code{npar} is \code{NULL}, so the degrees of freedom of
the chi-squared statistics are the number of classes minus 1.}

\item{grouping}{a logical indicating whether to group in classes with expected frequency greather or equal to 5. By default \code{grouping} is \code{FALSE}.}
}
\value{
A list with class \code{"htest"} containing the following components:
\itemize{
\item \code{statistic}: the value of the chi-squared test statistic.
\item \code{df}: the degrees of freedom of the approximate chi-squared distribution.
\item \code{p.value}: the p-value for the test.
\item \code{observed}: the observed counts.
\item \code{observed.grouped}: the observed counts grouped in classes with expected frequency greather or equal to 5.
\item \code{expected}: the expected counts under the null hypothesis.
\item \code{expected.grouped}: the expected counts under the null hypothesis grouped in classes with expected frequency greather or equal to 5.
\item \code{residuals}: the Pearson residuals computed as \code{(observed - expected) / sqrt(expected)}.
}
}
\description{
\code{chisq.test2} performs Pearson chi-squared goodness-of-fit test for count data
}
