% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dctp.R, R/pctp.R, R/qctp.R, R/rctp.R
\name{ebw}
\alias{ebw}
\alias{debw}
\alias{pebw}
\alias{qebw}
\alias{rebw}
\title{The Extended Biparametric Waring (EBW) Distribution COMPLETAR}
\usage{
debw(x, alpha, gamma, rho)

pebw(q, alpha, gamma, rho, lower.tail = TRUE)

qebw(p, alpha, gamma, rho, lower.tail = TRUE)

rebw(n, alpha, gamma, rho, lower.tail = TRUE)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles.}

\item{alpha}{parameter alpha (real)}

\item{gamma}{parameter \eqn{\gamma} (positive)}

\item{rho}{parameter rho (real)}

\item{q}{vector of quantiles.}

\item{lower.tail}{if TRUE (default), probabilities are \eqn{P(X<=x)}, otherwise, \eqn{P(X>x)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
\code{dctp} gives the pmf, \code{pctp} gives the distribution function, \code{qctp} gives the quantile function and \code{rctp} generates random values.

If \eqn{a = 0} the probability mass function, distribution function, quantile function and random generation function for the CBP distribution arise.
}
\description{
Probability mass function, distribution function, quantile function and random generation for the Complex Triparametric Pearson (CTP) distribution with parameters \eqn{a}, \eqn{b} and \eqn{\gamma}.
}
\details{
The CTP distribution with parameters \eqn{a}, \eqn{b} and \eqn{\gamma} has pmf
\deqn{f(x|a,b,\gamma) = C \Gamma(a+ib+x) \Gamma(a-ib+x) / (\Gamma(\gamma+x) x!), x=0,1,2,...} 
where \eqn{i} is the imaginary unit, \eqn{\Gamma(·)} the gamma function and 
\deqn{C = \Gamma(\gamma-a-ib) \Gamma(\gamma-a+ib) / (\Gamma(\gamma-2a) \Gamma(a+ib) \Gamma(a-ib))}
the normalizing constant.

If \eqn{a=0} the CTP is a Complex Biparametric Pearson (CBP) distribution, so the pmf of the CBP distribution is obtained.

The mean and the variance of the CTP distribution are
\eqn{E(X)=\mu=(a^2+b^2)/(\gamma-2a-1)} and \eqn{Var(X)=\mu(\mu+\gamma-1)/(\gamma-2a-2)}
so \eqn{\gamma > 2a + 2}.

It is underdispersed if \eqn{a < - (\mu + 1) / 2}, equidispersed if \eqn{a = - (\mu + 1) / 2} or overdispersed
if \eqn{a > - (\mu + 1) / 2}. In particular, if \eqn{a >= 0} the CTP is always overdispersed.
}
\examples{
# Examples for the function dctp
debw(3,1,rho=5)
debw(c(3,4),2,rho=5)

# Examples for the function pcbp
pebw(3,2,rho=5)
pebw(c(3,4),2,rho=5)

# Examples for the function pcbp
qebw(0.5,2,rho=5)
qebw(c(.8,.9),2,rho=5)

# Examples for the function rebw
rebw(4,2,rho=5)

}
\references{
\insertRef{RCS2003}{cpd}

\insertRef{RCSO2004}{cpd}

\insertRef{ROC2018}{cpd}
}
\seealso{
Functions for maximum-likelihood fitting of the CTP and CBP distributions: \code{\link{fitctp}} and \code{\link{fitcbp}}.
}
