% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitctp.R
\name{fitcbp}
\alias{fitcbp}
\title{Maximum likelihood estimation fit for CBP(b,gamma)}
\usage{
fitcbp(x, bstart = 1, gammastart = 1.1, method = "L-BFGS-B",
  moments = FALSE, hessian = TRUE, control = list(), ...)
}
\arguments{
\item{x}{A param x}

\item{bstart}{A param bstart}

\item{gammastart}{A param gammastart}

\item{method}{A param method ......}

\item{moments}{A param TRUE FALSE that say if we want to use initial values generating for moments method if posible.}

\item{hessian}{A param hessian TRUE FALSE}

\item{control}{A param control}

\item{...}{other params}
}
\value{
la salida del ...
}
\description{
Maximum likelihood estimation fit for CBP(b,gamma)
}
\examples{
x<-c(3,5,7,2,6,9,7,5)
fitcbp(x,1.1,3)

}
