% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitctp.R
\name{varcomp}
\alias{varcomp}
\title{varcomp for fitEBW object.}
\usage{
varcomp(object, ...)
}
\arguments{
\item{object}{An object of class \code{'fitebw'}}

\item{...}{Additional parameters.}
}
\value{
Two data frames, with ratio of sources of variation and sources of variation in which variance is splitted.
}
\description{
Description .
}
\examples{

set.seed(123)
x <- rebw(500, 2,rho=5)
fit<-fitebw(x, alphastart = 1, rhostart = 5)

varcomp(fit)
}
